Enable use of SATA port multiplier
This commit changes the kernel so it always expects a SATA port multiplier
to be connected to the eSATA port of the Banana Pi. A single hard disk will
no longer work after this change.

Connecting a port multiplier is practical if one wants to use the Banana Pi
as a NAS and connect two or more harddisks to create a RAID setup.